#46 Altera tipo de id long para a classe Long e corrige pontos que estavam faltando. Altera pacotes dos testes.

+++ src/main/java/br/ufg/inf/pitanga/interfaces/dao/InterfacePecaDao.java
@@ -15,5 +15,5 @@ public interface InterfacePecaDao {
 
     public void alterarPeca(Peca umaPeca);
 
-    public Peca buscarPorId(int id);
+    public Peca buscarPorId(Long id);
 }
